move environment codegen into plugin-utils
Any addon that wants to discover the basic hub environment configuration should do it by importing `@cardstack/plugin-utils/environment`. This replaces `@cardstack/hub/environment`, because plugins aren't supposed to directly depend on the hub package, which made the previous naming confusing.

I've also added a fallback so that if no hub is configured plugins can still import (default) configuration without throwing errors.
