+++ packages/authentication/package.json
@@ -26,6 +26,7 @@
   "dependencies": {
     "@cardstack/di": "^0.4.4",
     "@cardstack/logger": "^0.1.0",
+    "@cardstack/plugin-utils": "^0.4.4",
     "ember-cli-babel": "^6.8.2",
     "ember-cli-htmlbars": "^2.0.1",
     "ember-fetch": "^1.4.0",
@@ -38,7 +39,6 @@
   },
   "devDependencies": {
     "@cardstack/eslint-config": "^0.4.4",
-    "@cardstack/plugin-utils": "^0.4.4",
     "@cardstack/test-support": "^0.4.4",
     "broccoli-asset-rev": "^2.4.5",
     "ember-cli": "https://github.com/ember-cli/ember-cli#a4734540b3889c7f6ef304fe6a81dd356f77bfe4",
