+++ packages/codegen/package.json
@@ -31,6 +31,7 @@
   },
   "devDependencies": {
     "@cardstack/eslint-config": "^0.4.4",
+    "@cardstack/jsonapi": "^0.4.4",
     "@cardstack/test-support": "^0.4.4",
     "ember-cli": "https://github.com/ember-cli/ember-cli#a4734540b3889c7f6ef304fe6a81dd356f77bfe4",
     "ember-cli-dependency-checker": "^2.1.0",
