+++ packages/models/addon/adapter.js
@@ -1,7 +1,7 @@
 import DS from 'ember-data';
 import AdapterMixin from 'ember-resource-metadata/adapter-mixin';
 import Ember from 'ember';
-import { hubURL } from '@cardstack/hub/environment';
+import { hubURL } from '@cardstack/plugin-utils/environment';
 
 export default DS.JSONAPIAdapter.extend(AdapterMixin, {
   host: hubURL,
