+++ packages/models/package.json
@@ -25,6 +25,7 @@
   "dependencies": {
     "@cardstack/codegen": "^0.4.4",
     "@cardstack/di": "^0.4.4",
+    "@cardstack/plugin-utils": "^0.4.4",
     "@cardstack/tools": "^0.4.4",
     "ember-cli-babel": "^6.8.2",
     "ember-resource-metadata": "^0.1.0",
