+++ packages/plugin-utils/addon/.eslintrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  root: true,
+  "extends": require.resolve('@cardstack/eslint-config/browser')
+};
