+++ packages/plugin-utils/package.json
@@ -18,14 +18,19 @@
     "walk-sync": "^0.3.2"
   },
   "dependencies": {
-    "@cardstack/jsonapi": "^0.4.4",
     "broccoli-plugin": "^1.3.0",
+    "ember-cli-babel": "^6.8.2",
+    "handlebars": "^4.0.6",
     "ms": "2.0.0",
     "superagent": "^3.8.1",
     "symlink-or-copy": "^1.1.8"
   },
+  "peerDependencies": {
+    "@cardstack/jsonapi": "^0.4.4"
+  },
   "keywords": [
-    "cardstack-plugin"
+    "cardstack-plugin",
+    "ember-addon"
   ],
   "cardstack-plugin": {
     "api-version": 1
