+++ packages/tools/tests/dummy/app/adapters/post.js
@@ -1,6 +1,6 @@
 import AdapterMixin from 'ember-resource-metadata/adapter-mixin';
 import DS from 'ember-data';
-import { hubURL } from '@cardstack/hub/environment';
+import { hubURL } from '@cardstack/plugin-utils/environment';
 
 export default DS.JSONAPIAdapter.extend(AdapterMixin, {
   host: hubURL,
